int vs float credit_balance

+++ app/schemas/agent_controller.py
@@ -13,7 +13,7 @@ class OAuth2Request(BaseModel):
 class AgentControllerResponse(AgentControllerBase):
     id: str
     api_key: str = Field(..., description="API key for authentication")
-    credit_balance: int = Field(default=10, description="Credits. In cents of USD")
+    credit_balance: float = Field(default=10, description="Credits. In cents of USD")
     created_at: datetime
     last_login: Optional[datetime] = None
